requirements.txt
python-telegram-bot==13.15
python-dotenv==1.0.1
mysql-connector-python==8.0.32
requests==2.31.0
beautifulsoup4==4.12.2
